most updated version of biac app

+++ classified_image/models.py
@@ -3,7 +3,7 @@ from image.models import Image
 
 class Classified_image (models.Model) :
     image_id = models.ForeignKey(Image,related_name='classified_image' ,on_delete=models.CASCADE)
-    volunteers_evaluation_score = models.IntegerField(null=True)
+    volunteers_evaluation_score = models.FloatField(null=True)
     number_of_evalutors_per_image = models.IntegerField(null=True)
     image_with_model_classification = models.ImageField(null=False)
     confidence_score = models.FloatField(null=True)
